+++ scenedetect/frame_timecode.py
@@ -6,7 +6,7 @@
 #     [  Github: https://github.com/Breakthrough/PySceneDetect/  ]
 #     [  Documentation: http://pyscenedetect.readthedocs.org/    ]
 #
-# Copyright (C) 2014-2021 Brandon Castellano <http://www.bcastell.com>.
+# Copyright (C) 2014-2022 Brandon Castellano <http://www.bcastell.com>.
 #
 # PySceneDetect is licensed under the BSD 3-Clause License; see the included
 # LICENSE file, or visit one of the following pages for details:
@@ -20,8 +20,7 @@
 # ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
-
-""" ``scenedetect.frame_timecode`` Module
+"""``scenedetect.frame_timecode`` Module
 
 This module contains the :py:class:`FrameTimecode` object, which is used as a way for
 PySceneDetect to store frame-accurate timestamps of each cut. This is done by also
@@ -34,20 +33,18 @@ See the following examples, or the :py:class:`FrameTimecode constructor <FrameTi
 Unit tests for the FrameTimecode object can be found in `tests/test_timecode.py`.
 """
 
-# Standard Library Imports
 import math
+from typing import Union
+
+MAX_FPS_DELTA = 1.0 / 100000
+"""Maximum amount two framerates can differ by for equality testing."""
 
 
-MINIMUM_FRAMES_PER_SECOND_FLOAT = 1.0 / 1000.0
-MINIMUM_FRAMES_PER_SECOND_DELTA_FLOAT = 1.0 / 100000
+class FrameTimecode:
+    """Object for frame-based timecodes, using the video framerate to compute back and
+    forth between frame number and seconds/timecode.
 
-
-class FrameTimecode(object):
-    """ Object for frame-based timecodes, using the video framerate
-    to compute back and forth between frame number and second/timecode formats.
-
-    The timecode argument is valid only if it complies with one of the following
-    three types/formats:
+    A timecode is valid only if it complies with one of the following three types/formats:
 
     1) string: standard timecode HH:MM:SS[.nnn]:
         `str` in form 'HH:MM:SS' or 'HH:MM:SS.nnn', or
@@ -60,27 +57,23 @@ class FrameTimecode(object):
         `str` in form 'N'
 
     Arguments:
-        timecode (str, float, int, or FrameTimecode):  A timecode or frame
-            number, given in any of the above valid formats/types.  This
-            argument is always required.
-        fps (float, or FrameTimecode, conditionally required): The framerate
-            to base all frame to time arithmetic on (if FrameTimecode, copied
-            from the passed framerate), to allow frame-accurate arithmetic. The
-            framerate must be the same when combining FrameTimecode objects
-            in operations. This argument is always required, unless **timecode**
-            is a FrameTimecode.
+        timecode: A timecode or frame number, given in any of the above valid formats/types.
+            This argument is always required. If set to a FrameTimecode, it will be cloned.
+        fps: The framerate or another FrameTimecode to base all time arithmetic on.
     Raises:
         TypeError: Thrown if timecode is wrong type/format, or if fps is None
             or a type other than int or float.
         ValueError: Thrown when specifying a negative timecode or framerate.
     """
 
-    def __init__(self, timecode=None, fps=None):
-        # type: (Union[int, float, str, FrameTimecode], float,
-        #        Union[int, float, str, FrameTimecode])
+    def __init__(self,
+                 timecode: Union[int, float, str, 'FrameTimecode'] = None,
+                 fps: Union[int, float, str, 'FrameTimecode'] = None):
+
         # The following two properties are what is used to keep track of time
         # in a frame-specific manner.  Note that once the framerate is set,
         # the value should never be modified (only read if required).
+        # TODO(v1.0): Make these actual @properties.
         self.framerate = None
         self.frame_num = None
 
@@ -100,8 +93,8 @@ class FrameTimecode(object):
             # Process the given framerate, if it was not already set.
             if not isinstance(fps, (int, float)):
                 raise TypeError('Framerate must be of type int/float.')
-            elif (isinstance(fps, int) and not fps > 0) or (
-                    isinstance(fps, float) and not fps >= MINIMUM_FRAMES_PER_SECOND_FLOAT):
+            if (isinstance(fps, int) and not fps > 0) or (isinstance(fps, float)
+                                                          and not fps >= MAX_FPS_DELTA):
                 raise ValueError('Framerate must be positive and greater than zero.')
             self.framerate = float(fps)
 
@@ -111,23 +104,9 @@ class FrameTimecode(object):
         else:
             self.frame_num = self._parse_timecode_number(timecode)
 
-        # Alternative formats under consideration (require unit tests before adding):
-
-        # Standard timecode in list format [HH, MM, SS.nnn]
-        #elif isinstance(timecode, (list, tuple)) and len(timecode) == 3:
-        #    if any(not isinstance(x, (int, float)) for x in timecode):
-        #        raise ValueError('Timecode components must be of type int/float.')
-        #    hrs, mins, secs = timecode
-        #    if not (hrs >= 0 and mins >= 0 and secs >= 0 and mins < 60
-        #            and secs < 60):
-        #        raise ValueError('Timecode components must be positive.')
-        #    secs += (((hrs * 60.0) + mins) * 60.0)
-        #    self.frame_num = int(secs * self.framerate)
-
-
-    def get_frames(self):
-        # type: () -> int
-        """ Get the current time/position in number of frames.  This is the
+    # TODO(v1.0): Add a `frame` property to replace the existing one and deprecate this getter.
+    def get_frames(self) -> int:
+        """Get the current time/position in number of frames.  This is the
         equivalent of accessing the self.frame_num property (which, along
         with the specified framerate, forms the base for all of the other
         time measurement calculations, e.g. the :py:meth:`get_seconds` method).
@@ -140,36 +119,30 @@ class FrameTimecode(object):
         """
         return int(self.frame_num)
 
-
-    def get_framerate(self):
-        # type: () -> float
-        """ Get Framerate: Returns the framerate used by the FrameTimecode object.
+    # TODO(v1.0): Add a `framerate` property to replace the existing one and deprecate this getter.
+    def get_framerate(self) -> float:
+        """Get Framerate: Returns the framerate used by the FrameTimecode object.
 
         Returns:
             float: Framerate of the current FrameTimecode object, in frames per second.
         """
         return self.framerate
 
-
-    def equal_framerate(self, fps):
-        # type: (float) -> bool
-        """ Equal Framerate: Determines if the passed framerate is equal to that of the
-        FrameTimecode object.
+    def equal_framerate(self, fps) -> bool:
+        """Equal Framerate: Determines if the passed framerate is equal to that of this object.
 
         Arguments:
-            fps:    Framerate (float) to compare against within the precision constant
-                    MINIMUM_FRAMES_PER_SECOND_DELTA_FLOAT defined in this module.
-
+            fps: Framerate to compare against within the precision constant defined in this module
+                (see :py:data:`MAX_FPS_DELTA`).
         Returns:
             bool: True if passed fps matches the FrameTimecode object's framerate, False otherwise.
 
         """
-        return math.fabs(self.framerate - fps) < MINIMUM_FRAMES_PER_SECOND_DELTA_FLOAT
+        return math.fabs(self.framerate - fps) < MAX_FPS_DELTA
 
-
-    def get_seconds(self):
-        # type: () -> float
-        """ Get the frame's position in number of seconds.
+    # TODO(v1.0): Add a `seconds` property to replace this and deprecate the existing one.
+    def get_seconds(self) -> float:
+        """Get the frame's position in number of seconds.
 
         If using to compare a :py:class:`FrameTimecode` with a frame number,
         you can do so directly against the object (e.g. ``FrameTimecode(10, 10.0) <= 1.0``).
@@ -179,14 +152,14 @@ class FrameTimecode(object):
         """
         return float(self.frame_num) / self.framerate
 
-
-    def get_timecode(self, precision=3, use_rounding=True):
-        # type: (int, bool) -> str
-        """ Get a formatted timecode string of the form HH:MM:SS[.nnn].
+    # TODO(v1.0): Add a `timecode` property to replace this and deprecate the existing one.
+    def get_timecode(self, precision: int = 3, use_rounding: bool = True) -> str:
+        """Get a formatted timecode string of the form HH:MM:SS[.nnn].
 
         Args:
-            precision:     The number of decimal places to include in the output ``[.nnn]``.
-            use_rounding:  True (default) to round the output to the desired precision.
+            precision: The number of decimal places to include in the output ``[.nnn]``.
+            use_rounding: Rounds the output to the desired precision. If False, the value
+                will be truncated to the specified precision.
 
         Returns:
             str: The current time in the form ``"HH:MM:SS[.nnn]"``.
@@ -203,7 +176,6 @@ class FrameTimecode(object):
         if precision > 0:
             if use_rounding:
                 secs = round(secs, precision)
-                #secs = math.ceil(secs * (10**precision)) / float(10**precision)
             msec = format(secs, '.%df' % precision)[-precision:]
             secs = '%02d.%s' % (int(secs), msec)
         else:
@@ -211,19 +183,15 @@ class FrameTimecode(object):
         # Return hours, minutes, and seconds as a formatted timecode string.
         return '%02d:%02d:%s' % (hrs, mins, secs)
 
-    def previous_frame(self):
-        # type: () -> FrameTimecode
-        """
-        Returns a new FrameTimecode for the frame before this one.
-        :return: New FrameTimeCode object, one frame earlier
-        """
+    # TODO(v1.0): Add a `previous` property to replace the existing one and deprecate this getter.
+    def previous_frame(self) -> 'FrameTimecode':
+        """Return a new FrameTimecode for the previous frame (or 0 if on frame 0)."""
         new_timecode = FrameTimecode(self)
-        new_timecode.frame_num -= 1
+        new_timecode.frame_num = max(0, new_timecode.frame_num - 1)
         return new_timecode
 
-    def _seconds_to_frames(self, seconds):
-        # type: (float) -> int
-        """ Converts the passed value seconds to the nearest number of frames using
+    def _seconds_to_frames(self, seconds: float) -> int:
+        """Convert the passed value seconds to the nearest number of frames using
         the current FrameTimecode object's FPS (self.framerate).
 
         Returns:
@@ -232,10 +200,8 @@ class FrameTimecode(object):
         """
         return int(seconds * self.framerate)
 
-
-    def _parse_timecode_number(self, timecode):
-        # type: (Union[int, float]) -> int
-        """ Parses a timecode number, storing it as the exact number of frames.
+    def _parse_timecode_number(self, timecode: Union[int, float]) -> int:
+        """ Parse a timecode number, storing it as the exact number of frames.
         Can be passed as frame number (int), seconds (float)
 
         Raises:
@@ -260,12 +226,11 @@ class FrameTimecode(object):
         else:
             raise TypeError('Timecode format/type unrecognized.')
 
-
-    def _parse_timecode_string(self, timecode_string):
-        # type: (str) -> int
-        """ Parses a string based on the three possible forms (in timecode format,
+    def _parse_timecode_string(self, timecode_string: str) -> int:
+        """Parses a string based on the three possible forms (in timecode format,
         as an integer number of frames, or floating-point seconds, ending with 's').
-        Requires that the framerate property is set before calling this method.
+
+        Requires that the `framerate` property is set before calling this method.
         Assuming a framerate of 30.0 FPS, the strings '00:05:00.000', '00:05:00',
         '9000', '300s', and '300.0s' are all possible valid values, all representing
         a period of time equal to 5 minutes, 300 seconds, or 9000 frames (at 30 FPS).
@@ -303,9 +268,7 @@ class FrameTimecode(object):
             secs += (((hrs * 60.0) + mins) * 60.0)
             return int(secs * self.framerate)
 
-
-    def __iadd__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> FrameTimecode
+    def __iadd__(self, other: Union[int, float, str, 'FrameTimecode']) -> 'FrameTimecode':
         if isinstance(other, int):
             self.frame_num += other
         elif isinstance(other, FrameTimecode):
@@ -318,20 +281,16 @@ class FrameTimecode(object):
             self.frame_num += self._seconds_to_frames(other)
         else:
             raise TypeError('Unsupported type for performing addition with FrameTimecode.')
-        if self.frame_num < 0:     # Required to allow adding negative seconds/frames.
+        if self.frame_num < 0: # Required to allow adding negative seconds/frames.
             self.frame_num = 0
         return self
 
-
-    def __add__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> FrameTimecode
+    def __add__(self, other: Union[int, float, str, 'FrameTimecode']) -> 'FrameTimecode':
         to_return = FrameTimecode(timecode=self)
         to_return += other
         return to_return
 
-
-    def __isub__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> FrameTimecode
+    def __isub__(self, other: Union[int, float, str, 'FrameTimecode']) -> 'FrameTimecode':
         if isinstance(other, int):
             self.frame_num -= other
         elif isinstance(other, FrameTimecode):
@@ -343,21 +302,18 @@ class FrameTimecode(object):
         elif isinstance(other, float):
             self.frame_num -= self._seconds_to_frames(other)
         else:
-            raise TypeError('Unsupported type for performing subtraction with FrameTimecode: %s' % type(other))
+            raise TypeError('Unsupported type for performing subtraction with FrameTimecode: %s' %
+                            type(other))
         if self.frame_num < 0:
             self.frame_num = 0
         return self
 
-
-    def __sub__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> FrameTimecode
+    def __sub__(self, other: Union[int, float, str, 'FrameTimecode']) -> 'FrameTimecode':
         to_return = FrameTimecode(timecode=self)
         to_return -= other
         return to_return
 
-
-    def __eq__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> bool
+    def __eq__(self, other: Union[int, float, str, 'FrameTimecode']) -> 'FrameTimecode':
         if isinstance(other, int):
             return self.frame_num == other
         elif isinstance(other, float):
@@ -373,16 +329,13 @@ class FrameTimecode(object):
         elif other is None:
             return False
         else:
-            raise TypeError('Unsupported type for performing == with FrameTimecode: %s' % type(other))
+            raise TypeError('Unsupported type for performing == with FrameTimecode: %s' %
+                            type(other))
 
-
-    def __ne__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> bool
+    def __ne__(self, other: Union[int, float, str, 'FrameTimecode']) -> bool:
         return not self == other
 
-
-    def __lt__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> bool
+    def __lt__(self, other: Union[int, float, str, 'FrameTimecode']) -> bool:
         if isinstance(other, int):
             return self.frame_num < other
         elif isinstance(other, float):
@@ -395,14 +348,11 @@ class FrameTimecode(object):
             else:
                 raise TypeError(
                     'FrameTimecode objects must have the same framerate to be compared.')
-        #elif other is None:
-        #    return False
         else:
-            raise TypeError('Unsupported type for performing < with FrameTimecode: %s' % type(other))
+            raise TypeError('Unsupported type for performing < with FrameTimecode: %s' %
+                            type(other))
 
-
-    def __le__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> bool
+    def __le__(self, other: Union[int, float, str, 'FrameTimecode']) -> bool:
         if isinstance(other, int):
             return self.frame_num <= other
         elif isinstance(other, float):
@@ -415,14 +365,11 @@ class FrameTimecode(object):
             else:
                 raise TypeError(
                     'FrameTimecode objects must have the same framerate to be compared.')
-        #elif other is None:
-        #    return False
         else:
-            raise TypeError('Unsupported type for performing <= with FrameTimecode: %s' % type(other))
+            raise TypeError('Unsupported type for performing <= with FrameTimecode: %s' %
+                            type(other))
 
-
-    def __gt__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> bool
+    def __gt__(self, other: Union[int, float, str, 'FrameTimecode']) -> bool:
         if isinstance(other, int):
             return self.frame_num > other
         elif isinstance(other, float):
@@ -435,15 +382,11 @@ class FrameTimecode(object):
             else:
                 raise TypeError(
                     'FrameTimecode objects must have the same framerate to be compared.')
-        #elif other is None:
-        #    return False
         else:
             raise TypeError('Unsupported type for performing > with FrameTimecode: %s' %
                             type(other))
 
-
-    def __ge__(self, other):
-        # type: (Union[int, float, str, FrameTimecode]) -> bool
+    def __ge__(self, other: Union[int, float, str, 'FrameTimecode']) -> bool:
         if isinstance(other, int):
             return self.frame_num >= other
         elif isinstance(other, float):
@@ -456,25 +399,24 @@ class FrameTimecode(object):
             else:
                 raise TypeError(
                     'FrameTimecode objects must have the same framerate to be compared.')
-        #elif other is None:
-        #    return False
         else:
-            raise TypeError('Unsupported type for performing >= with FrameTimecode: %s' % type(other))
+            raise TypeError('Unsupported type for performing >= with FrameTimecode: %s' %
+                            type(other))
 
+    # TODO(v1.0): __int__ and __float__ should be removed. Mark as deprecated, and indicate
+    # need to use relevant property instead.
 
-    # TODO: __int__ and __float__ should be removed.
-
-    def __int__(self):
+    def __int__(self) -> int:
         return self.frame_num
 
-    def __float__(self):
+    def __float__(self) -> float:
         return self.get_seconds()
 
-    def __str__(self):
+    def __str__(self) -> str:
         return self.get_timecode()
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return '%s [frame=%d, fps=%.3f]' % (self.get_timecode(), self.frame_num, self.framerate)
 
-    def __hash__(self):
+    def __hash__(self) -> int:
         return self.frame_num